requirements.txt
cachelib==0.1
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
Flask==1.1.2
Flask-Session==0.3.2
Flask-WeasyPrint==0.6
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
Pyphen==0.9.5
python-dotenv==0.13.0
redis==3.5.3
six==1.15.0
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
